Base controller user ID is now fetched on property value access, not in the controller constructor
This fixes a bug where if the user was re-authenticated by a persistent session cookie in the auth filter, the dashboard would be blank until it was refreshed again. This is because the controller user ID variable was still set to -1 (because that's what the session contained when the controller was instantiated) and it was never re-set. This way is much safer because it always reads the *current* value of the session cookie.